move code to the homepage and instructions to Readme.
